Fix splice_with_fallback flags type mismatch on Linux
libc::splice() expects flags as c_uint (u32) on Linux, not c_int (i32).
Pass the flags parameter directly instead of casting.

+++ src/socket_opts.rs
@@ -812,7 +812,7 @@ pub mod io_uring_splice {
                 pipe_w,
                 std::ptr::null_mut(),
                 len,
-                flags as libc::c_int,
+                flags,
             )
         };
         if n < 0 {
@@ -833,7 +833,7 @@ pub mod io_uring_splice {
                     fd_out,
                     std::ptr::null_mut(),
                     remaining,
-                    flags as libc::c_int,
+                    flags,
                 )
             };
             if written < 0 {
